can eval precision at l5
